+++ src/ASTParserFacade.ts
@@ -1,4 +1,5 @@
 import * as acorn from 'acorn';
+import acornImportMeta from 'acorn-import-meta';
 import * as ESTree from 'estree';
 
 import chalk, { Chalk } from 'chalk';
@@ -70,7 +71,11 @@ export class ASTParserFacade {
             sourceType
         };
 
-        const program: ESTree.Program = <any>acorn.parse(input, config);
+        const program: ESTree.Program = (
+            <any>acorn
+                .Parser.extend(acornImportMeta)
+                .parse(input, config)
+        );
 
         if (comments.length) {
             program.comments = comments;