+++ src/declarations/acorn-import-meta.d.ts
@@ -1,0 +1,7 @@
+declare module 'acorn-import-meta' {
+    import * as acorn from 'acorn';
+
+    function acornImportMeta (BaseParser: typeof acorn.Parser): typeof acorn.Parser;
+
+    export = acornImportMeta;
+}